add ruby version

+++ app/controllers/attachments_controller.rb
@@ -1,8 +1,8 @@
 # app/controllers/attachments_controller.rb
 class AttachmentsController < ApplicationController
   def preview
-    file_name = "#{params[:filename]}.#{params[:format]}"
-    file_path = Rails.root.join("public", "uploads", "attachments", "41", file_name)
+    file_name = File.basename("#{params[:filename]}.#{params[:format]}")
+    file_path = Rails.root.join("public/uploads/attachments/41/#{file_name}")
 
     if File.exist?(file_path)
       file_url = "#{request.protocol}#{request.host_with_port}/uploads/attachments/41/#{file_name}"
@@ -15,10 +15,9 @@ class AttachmentsController < ApplicationController
   def destroy
     attachment = Attachment.find_by(id: params[:id])
     if attachment
-      file_path = Rails.root.join("public", "uploads", "attachments", "41", attachment.file_name)
-      if File.exist?(file_path)
-        File.delete(file_path)
-      end
+      file_name = File.basename(attachment.file_name)
+      file_path = Rails.root.join("public/uploads/attachments/41/#{file_name}")
+      File.delete(file_path) if File.exist?(file_path)
       attachment.destroy
       render json: { message: "Attachment deleted successfully" }, status: :ok
     else
